BREL-37 #comment Apply Type casting to fix VAT_RATE_AFTER_DISCOUNT issue

+++ component/site/helpers/cart.php
@@ -89,7 +89,7 @@ class rsCarthelper
 
 				if (strstr($data, '{tax_after_discount}'))
 				{
-					if (APPLY_VAT_ON_DISCOUNT && VAT_RATE_AFTER_DISCOUNT)
+					if (APPLY_VAT_ON_DISCOUNT && (float) VAT_RATE_AFTER_DISCOUNT)
 					{
 						if ($check)
 						{
@@ -140,11 +140,11 @@ class rsCarthelper
 		$tax_after_discount = 0;
 		$cart               = $this->_session->get('cart');
 
-		if (APPLY_VAT_ON_DISCOUNT && VAT_RATE_AFTER_DISCOUNT)
+		if (APPLY_VAT_ON_DISCOUNT && (float) VAT_RATE_AFTER_DISCOUNT)
 		{
 			if ($discount > 0)
 			{
-				$tmptax             = VAT_RATE_AFTER_DISCOUNT * $discount;
+				$tmptax             = (float) VAT_RATE_AFTER_DISCOUNT * $discount;
 				$tax_after_discount = $tax - $tmptax;
 			}
 		}
@@ -2307,7 +2307,7 @@ class rsCarthelper
 
 		$chktag = $this->_producthelper->taxexempt_addtocart();
 
-		if (VAT_RATE_AFTER_DISCOUNT && !APPLY_VAT_ON_DISCOUNT && !empty($chktag))
+		if ((float) VAT_RATE_AFTER_DISCOUNT && !APPLY_VAT_ON_DISCOUNT && !empty($chktag))
 		{
 			if (isset($cart['discount_tax']) && !empty($cart['discount_tax']))
 			{
@@ -4221,7 +4221,7 @@ class rsCarthelper
 				{
 					$avgVAT = 1;
 
-					if (VAT_RATE_AFTER_DISCOUNT && !APPLY_VAT_ON_DISCOUNT)
+					if ((float) VAT_RATE_AFTER_DISCOUNT && !APPLY_VAT_ON_DISCOUNT)
 					{
 						$productVAT = $cart['product_subtotal'] - $cart['product_subtotal_excl_vat'];
 						$avgVAT = $cart['product_subtotal'] / $cart['product_subtotal_excl_vat'];
@@ -4793,7 +4793,7 @@ class rsCarthelper
 		$Discountvat = 0;
 		$chktag      = $this->_producthelper->taxexempt_addtocart();
 
-		if (VAT_RATE_AFTER_DISCOUNT && !empty($chktag) && !APPLY_VAT_ON_DISCOUNT)
+		if ((float) VAT_RATE_AFTER_DISCOUNT && !empty($chktag) && !APPLY_VAT_ON_DISCOUNT)
 		{
 			$vatData = $this->_producthelper->getVatRates();
